+++ deltafi-core/src/main/java/org/deltafi/core/configuration/DeltaFiProperties.java
@@ -26,6 +26,7 @@ public class DeltaFiProperties {
 
     private int requeueSeconds = 300;
     private int coreServiceThreads = 16;
+    private int scheduledServiceThreads = 32;
     private DeleteProperties delete = new DeleteProperties();
     private IngressProperties ingress = new IngressProperties();
     private String apiUrl;